requirements.txt
django == 1.7
django-social-auth
oauth2client
django-allauth
importlib